Fix dependencies for a lot of xorg packages.
git-svn-id: svn://svn.openwrt.org/openwrt@12522 3c298f89-4303-0410-b956-a3cf2f4a3e73

+++ packages/XOrg/generate.sh
@@ -10,14 +10,18 @@ for d in `find . -name packages`; do
 		DEP=`echo ${p##*|} | sed "s/+/ +/g"`
 		echo generating Makefile for ${NAME}-${VER} with deps : ${DEP}
 		rm -f ${NAME}/Makefile
-        rm -f ${NAME}/patches/*
+        if [ -e ${NAME}/patches ]; then
+            rm -f ${NAME}/patches/*
+        fi
 		if [ "$1" = "gen" ]; then
 			if [ ! -e ${NAME} ]; then
                 mkdir ${NAME}
             fi
 			sed "s/@VER@/${VER}/g" template.mk | sed "s/@DEP@/${DEP}/g" | sed "s/@NAME@/${NAME}/g" | sed "s/@BASE_NAME@/${BASE_NAME}/g" > ${NAME}/Makefile
 			if [ -d `pwd`/patches/${NAME} ]; then
-				mkdir ${NAME}/patches
+				if [ ! -d ${NAME}/patches ]; then
+                    mkdir ${NAME}/patches
+                fi
 				cp -r `pwd`/patches/${NAME}/* ${NAME}/patches/
 			fi
 		fi
